backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
alembic>=1.11.0
passlib[bcrypt]>=1.7.0
python-jose>=3.3.0
boto3>=1.26.0
httpx>=0.24.0
pytest>=7.0.0
email-validator>=1.3.0
